backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Configuration & Environment
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Async Support
aiofiles==23.2.1

# Database - PostgreSQL (async)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Vector Database - Qdrant
qdrant-client==1.7.3

# Object Storage - MinIO
minio==7.2.3

# ML - CLIP Embeddings
open-clip-torch==2.24.0
Pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0

# Logging
python-json-logger==2.0.7

# Code Quality (linting/formatting - optional but recommended)
black==23.12.1
ruff==0.1.11